refactor: move unit tests to appropriate directory
